chore(app): analysis information move
